requirements.txt
--extra-index-url https://${ARTIFACTORY_SERVER}/artifactory/api/pypi/f5-cloud-solutions-pypi/simple

# production packages
click==7.0
pyyaml==5.3.*
click-repl==0.1.*
f5-sdk-python==0.9.*
f5-teem==1.0.*
# doc packages
f5-sphinx-theme==2.0.5
sphinx==1.8.5
sphinxjp.themes.basicstrap==0.5.0
cloud_sptheme==1.9.4
recommonmark==0.6.0
sphinx_copybutton==0.2.5
# dev packages
pytest==5.2.2
pytest-cov==2.8.1
pytest-mock==1.11.2
flake8==3.7.9
pylint==2.3.1 ; python_version > '3.0'
safety==1.8.5
# build packages
twine==3.* ; python_version >= '3.0'
twine==1.* ; python_version <= '3.0'
